Refactor to allow for secure overrides
Uses a hardcoded path, however this is OK! since the path referrs
to one in the docker container which we fully control.

Note, that this also contains a is file when clause so we're only
executing this task for services that have a valid override

ensures the override directory is created before trying to copy the j2
into place. Additionally uses hardcoded UID/GID. However we know docker
will always mount the user as 1000:1000